bpf: IPv4 forwarding & connection tracking
Splits CT into IPv4 and IPv6 map. Transparent to user. Adds a second
entry to lxcmap. Forwarding functional. Sending of ICMP TTL exceeded
not implemented yet.

Acked-by: André Martins <andre@cilium.io>

+++ bpf/lib/maps.h
@@ -7,7 +7,7 @@
 #define CILIUM_MAP_POLICY	1
 #define CILIUM_MAP_CALLS	2
 
-__BPF_MAP(cilium_lxc, BPF_MAP_TYPE_HASH, CILIUM_MAP_LXC, sizeof(__u16), sizeof(struct lxc_info), PIN_GLOBAL_NS, 1024);
+__BPF_MAP(cilium_lxc, BPF_MAP_TYPE_HASH, CILIUM_MAP_LXC, sizeof(__u32), sizeof(struct lxc_info), PIN_GLOBAL_NS, 1024);
 
 /* Global map to jump into policy enforcement of receiving endpoint */
 BPF_PROG_ARRAY(cilium_policy, CILIUM_MAP_POLICY, PIN_GLOBAL_NS, POLICY_MAP_SIZE);
